backend/python-service/plate_recognition/requirements.txt
flask==3.1.2
flask-cors==6.0.2
ultralytics==8.3.241
torch==2.2.2
torchvision==0.17.2
onnxruntime==1.19.2
opencv-python==4.12.0.88
numpy==1.26.4
pillow==11.3.0
pyyaml==6.0.3
